mission-alpha-drone/backend/requirements.txt
google-genai
websockets
python-dotenv
fastapi
uvicorn
